move vestiges of shared account mod to casper-types.
