+++ execution_engine/src/shared.rs
@@ -3,7 +3,6 @@
 pub mod additive_map;
 #[macro_use]
 pub mod gas;
-pub mod account;
 pub mod host_function_costs;
 pub mod logging;
 pub mod motes;
